fix: bad logic at header selector component
